* commit '39ff8149466ee8e86c0336f706cd4dd3713f359f':
  libstagefright: fix possible overflow in SampleTable.cpp

+++ media/libstagefright/SampleTable.cpp
@@ -336,7 +336,7 @@ status_t SampleTable::setTimeToSampleParams(
 
     mTimeToSampleCount = U32_AT(&header[4]);
     uint64_t allocSize = (uint64_t)mTimeToSampleCount * 2 * sizeof(uint32_t);
-    if (allocSize > SIZE_MAX) {
+    if (allocSize > UINT32_MAX) {
         return ERROR_OUT_OF_RANGE;
     }
     mTimeToSample = new (std::nothrow) uint32_t[mTimeToSampleCount * 2];
@@ -384,7 +384,7 @@ status_t SampleTable::setCompositionTimeToSampleParams(
 
     mNumCompositionTimeDeltaEntries = numEntries;
     uint64_t allocSize = (uint64_t)numEntries * 2 * sizeof(uint32_t);
-    if (allocSize > SIZE_MAX) {
+    if (allocSize > UINT32_MAX) {
         return ERROR_OUT_OF_RANGE;
     }
